Add direct dependency on Selenium to enforce version control
With no direct dependency on Selenium the Selenium BOM import in
Universe:Drone seems to be ignored when resolved along side the
Drone BOM which imports the Selenium BOM in a different version.

The dependencies seems correct when viewed from Universe:Drone, but
end up wrong when used by 3. party.